Moved FileToVoxCore into a project folder.
