+++ FileToVoxCore.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31919.166
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileToVoxCore", "FileToVoxCore.csproj", "{9E6D120D-23BD-41FC-B1D2-D64DD7EE4F40}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FileToVoxCore", "FileToVoxCore\FileToVoxCore.csproj", "{9B38A105-6687-42E7-AD6B-EE6CB844B137}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9E6D120D-23BD-41FC-B1D2-D64DD7EE4F40}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9E6D120D-23BD-41FC-B1D2-D64DD7EE4F40}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9E6D120D-23BD-41FC-B1D2-D64DD7EE4F40}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9E6D120D-23BD-41FC-B1D2-D64DD7EE4F40}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9B38A105-6687-42E7-AD6B-EE6CB844B137}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9B38A105-6687-42E7-AD6B-EE6CB844B137}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9B38A105-6687-42E7-AD6B-EE6CB844B137}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9B38A105-6687-42E7-AD6B-EE6CB844B137}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
