+++ FileToVoxCore/FileToVoxCore.csproj
@@ -1,0 +1,21 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Library</OutputType>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>FileToVoxCore</AssemblyName>
+    <RootNamespace>FileToVoxCore</RootNamespace>
+    <Version>1.3.1</Version>
+    <Authors>Zarbuz</Authors>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <PackageProjectUrl>https://github.com/Zarbuz/FileToVoxCore</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/Zarbuz/FileToVoxCore</RepositoryUrl>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Update="Schematics\config.txt">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+
+</Project>
