+++ FileToVoxCore/Schematics/Voxel.cs
