+++ FileToVoxCore/Vox/Chunks/GroupNodeChunk.cs
@@ -1,0 +1,8 @@
+﻿namespace FileToVoxCore.Vox.Chunks
+{
+    public class GroupNodeChunk : NodeChunk
+    { // nGRP: Group Node Chunk
+        public int[] ChildIds;
+        public override NodeType Type => NodeType.Group;
+    }
+}
