+++ FileToVoxCore/Vox/Chunks/ShapeNodeChunk.cs
@@ -1,0 +1,16 @@
+﻿using System.Collections.Generic;
+
+namespace FileToVoxCore.Vox.Chunks
+{
+    public class ShapeModel
+    {
+        public int ModelId;
+        public Dictionary<string, string> Attributes;
+    }
+
+    public class ShapeNodeChunk : NodeChunk
+    { // nSHP: Shape Node Chunk
+        public ShapeModel[] Models;
+        public override NodeType Type => NodeType.Shape;
+    }
+}
