--- Schematics/RawSchematic.cs
@@ -1,11 +1,0 @@
-﻿namespace FileToVoxCore.Schematics
-{
-    public class RawSchematic
-    {
-        public short Width;
-        public short Heigth;
-        public short Length;
-        public byte[] Blocks;
-        public byte[] Data;
-    }
-}
